requirements.txt
Flask==0.11.1
cf-deployment-tracker==1.0.4
python-dotenv==0.6.4
watson_developer_cloud==0.26.0
